Create tokens file
There were a lot of tokens so I created a different file for them
